Ft registry pattern planner (#24)
Introduces the registry pattern for planners.
Now, planners are created based on their name. However, the if-clause must no longer be manually added in the runner.
This is taken care of by the `PlannerRegistry'.
When a new planner is added in the future, it must derive from `Planner` and added to the `__init__.py` file in the planner folder.